# Conflicts:
#	envi/bits.py
#	envi/memcanvas/__init__.py
#	vivisect/analysis/__init__.py
#	vivisect/defconfig.py
#	vivisect/parsers/__init__.py
#	vivisect/parsers/macho.py
#	vivisect/tests/linux_amd64_ls_data.py
#	vivisect/tests/linux_i386_libstdc_data.py
#	vivisect/tests/openbsd_amd64_ls_data.py
#	vivisect/vivbin.py
#	vqt/hotkeys.py

+++ vdb/renderers.py
@@ -134,7 +134,7 @@ class SymbolRenderer(e_canvas.MemoryRenderer):
         if isptr:
             sym = trace.getSymByAddr(p, exact=False)
             if sym is not None:
-                mcanv.addText(' %s + %d' % (repr(sym), p-long(sym)))
+                mcanv.addText(' %s + %d' % (repr(sym), p-int(sym)))
 
         mcanv.addText('\n')
